+++ apps/backend/src/post/post-mutation.service.spec.ts
@@ -3,13 +3,12 @@ import { getDataSourceToken, getRepositoryToken } from "@nestjs/typeorm";
 import { DataSource, Repository } from "typeorm";
 import PGMem from "pg-mem";
 import { randomUUID } from "crypto";
-import { PostService } from "./post.service";
+import { PostMutationService } from "./post-mutation.service";
 import {
   AttachmentAlreadyAssociatedError,
   AttachmentNotFoundError,
   NotPostAuthorError,
   PostNotFoundError,
-  InvalidCursorError,
 } from "./post.errors";
 import { Post, PostCategory } from "./post.entity";
 import { PostLike } from "./post-like.entity";
@@ -69,9 +68,9 @@ async function saveTestAttachment(
   );
 }
 
-describe("PostService", () => {
+describe("PostMutationService", () => {
   let module: TestingModule;
-  let postService: PostService;
+  let postMutationService: PostMutationService;
   let postRepository: Repository<Post>;
   let postLikeRepository: Repository<PostLike>;
   let attachmentRepository: Repository<Attachment>;
@@ -89,7 +88,7 @@ describe("PostService", () => {
 
     module = await Test.createTestingModule({
       providers: [
-        PostService,
+        PostMutationService,
         {
           provide: getRepositoryToken(Post),
           useValue: dataSource.getRepository(Post),
@@ -110,7 +109,7 @@ describe("PostService", () => {
       ],
     }).compile();
 
-    postService = module.get(PostService);
+    postMutationService = module.get(PostMutationService);
     postRepository = dataSource.getRepository(Post);
     postLikeRepository = dataSource.getRepository(PostLike);
     attachmentRepository = dataSource.getRepository(Attachment);
@@ -136,7 +135,7 @@ describe("PostService", () => {
       const a2 = await saveTestAttachment(attachmentRepository);
 
       // When: 글 생성 시도
-      const post = await postService.createPost({
+      const post = await postMutationService.createPost({
         userId: user.id,
         content: "hello",
         nickname: "nick",
@@ -161,7 +160,7 @@ describe("PostService", () => {
       const nonExistentAttachmentId = 0;
 
       // When: 글 생성 시도
-      const result = postService.createPost({
+      const result = postMutationService.createPost({
         userId: user.id,
         content: "hello",
         nickname: "nick",
@@ -181,7 +180,7 @@ describe("PostService", () => {
       });
 
       // When: 글 생성 시도
-      const result = postService.createPost({
+      const result = postMutationService.createPost({
         userId: user.id,
         content: "hello",
         nickname: "nick",
@@ -203,7 +202,7 @@ describe("PostService", () => {
       });
 
       // When: 글 생성 시도
-      const result = postService.createPost({
+      const result = postMutationService.createPost({
         userId: user.id,
         content: "hello",
         nickname: "nick",
@@ -214,144 +213,6 @@ describe("PostService", () => {
       // Then: throws handled error, 글 생성되지 않음
       await expect(result).rejects.toThrow(AttachmentAlreadyAssociatedError);
       expect(await postRepository.count()).toBe(1); // only the existing post
-    });
-  });
-
-  describe("listPosts", () => {
-    // 2026-03-23: success 처리를 parameter의 조합별로 더 테스트하는 것이 더
-    // 엄밀하지만, 현재는 그럴 필요가 없다고 판단하여 success 케이스 하나만
-    // 처리합니다.
-    it("success: 각 parameter 처리", async () => {
-      // Given: posts
-      const user = await saveTestUser(userRepository);
-      const p1 = await saveTestPost(postRepository, user.id, {
-        category: PostCategory.DUMMY1,
-      });
-      const p2 = await saveTestPost(postRepository, user.id, {
-        category: PostCategory.DUMMY1,
-      });
-      const p3 = await saveTestPost(postRepository, user.id, {
-        category: PostCategory.DUMMY1,
-      });
-      await saveTestPost(postRepository, user.id, {
-        category: PostCategory.DUMMY2,
-      });
-
-      // When: 글 조회 시도
-      const result = await postService.listPosts({
-        limit: 2,
-        category: PostCategory.DUMMY1,
-        orderBy: "createdAt",
-        orderDirection: "asc",
-      });
-
-      // Then: 올바른 결과 반환
-      expect(result.posts.map((p) => p.id)).toEqual([p1.id, p2.id]);
-      expect(result.nextCursor).toBeDefined();
-
-      // When: nextCursor로 글 조회 시도
-      const result2 = await postService.listPosts({
-        cursor: result.nextCursor,
-        limit: 2,
-        category: PostCategory.DUMMY1,
-        orderBy: "createdAt",
-        orderDirection: "asc",
-      });
-
-      // Then: 올바른 결과 반환
-      expect(result2.posts.map((p) => p.id)).toEqual([p3.id]);
-      expect(result2.nextCursor).toBeUndefined();
-    });
-
-    it("success: cursor에 해당하는 글이 삭제된 경우 올바르게 처리", async () => {
-      // Given: cursor에 해당하는 글이 삭제된 경우
-      const user = await saveTestUser(userRepository);
-      const p1 = await saveTestPost(postRepository, user.id);
-      const p2 = await saveTestPost(postRepository, user.id);
-      const p3 = await saveTestPost(postRepository, user.id);
-
-      const firstPage = await postService.listPosts({
-        limit: 1,
-        orderBy: "createdAt",
-        orderDirection: "asc",
-      });
-      await postRepository.delete(p1.id);
-
-      // When: 글 조회 시도
-      const result = await postService.listPosts({
-        cursor: firstPage.nextCursor,
-        limit: 2,
-        orderBy: "createdAt",
-        orderDirection: "asc",
-      });
-
-      // Then: 삭제된 글 다음부터 올바르게 반환
-      expect(result.posts.map((p) => p.id)).toEqual([p2.id, p3.id]);
-    });
-
-    it("category 및 orderBy와 벗어난 cursor 처리", async () => {
-      // Given: cursor가 category 및 orderBy를 다르게 하여 조회했을 때의
-      // 결과물인 경우
-      const user = await saveTestUser(userRepository);
-      await saveTestPost(postRepository, user.id, {
-        category: PostCategory.DUMMY1,
-      });
-      await saveTestPost(postRepository, user.id, {
-        category: PostCategory.DUMMY1,
-      });
-      const firstPage = await postService.listPosts({
-        limit: 1,
-        orderBy: "createdAt",
-        orderDirection: "asc",
-        category: PostCategory.DUMMY1,
-      });
-
-      // When: 글 조회 시도
-      const result = postService.listPosts({
-        cursor: firstPage.nextCursor,
-        limit: 1,
-        orderBy: "createdAt",
-        orderDirection: "asc",
-        category: PostCategory.DUMMY2,
-      });
-
-      // Then: throws handled error
-      await expect(result).rejects.toThrow(InvalidCursorError);
-    });
-  });
-
-  describe("getPost", () => {
-    it("success: 올바른 글 반환", async () => {
-      // Given: 글이 존재하는 경우
-      const user = await saveTestUser(userRepository);
-      const post = await postRepository.save(
-        postRepository.create({
-          content: "hello world",
-          category: PostCategory.DUMMY1,
-          nickname: "testnick",
-          author: user,
-        }),
-      );
-
-      // When: 글 조회 시도
-      const result = await postService.getPost(post.id);
-
-      // Then: 올바른 글 반환
-      expect(result).toMatchObject({
-        id: post.id,
-        content: "hello world",
-        category: PostCategory.DUMMY1,
-      });
-    });
-
-    it("존재하지 않는 글 처리", async () => {
-      // Given: 글이 존재하지 않는 경우
-
-      // When: 글 조회 시도
-      const result = postService.getPost(0);
-
-      // Then: throws handled error
-      await expect(result).rejects.toThrow(PostNotFoundError);
     });
   });
 
@@ -369,7 +230,7 @@ describe("PostService", () => {
       const newAttachment = await saveTestAttachment(attachmentRepository);
 
       // When: 글 업데이트 시도
-      await postService.updatePost({
+      await postMutationService.updatePost({
         userId: user.id,
         postId: post.id,
         content: "updated content",
@@ -401,7 +262,7 @@ describe("PostService", () => {
       // Given: 글이 존재하지 않는 경우
 
       // When: 글 업데이트 시도
-      const result = postService.updatePost({
+      const result = postMutationService.updatePost({
         userId: 0,
         postId: 0,
         content: "content",
@@ -420,7 +281,7 @@ describe("PostService", () => {
       const post = await saveTestPost(postRepository, author.id);
 
       // When: 글 업데이트 시도
-      const result = postService.updatePost({
+      const result = postMutationService.updatePost({
         userId: otherUser.id,
         postId: post.id,
         content: "content",
@@ -441,7 +302,7 @@ describe("PostService", () => {
       const nonExistentAttachmentId = 0;
 
       // When: 글 업데이트 시도
-      const result = postService.updatePost({
+      const result = postMutationService.updatePost({
         userId: user.id,
         postId: post.id,
         content: "content",
@@ -464,7 +325,7 @@ describe("PostService", () => {
       });
 
       // When: 글 업데이트 시도
-      const result = postService.updatePost({
+      const result = postMutationService.updatePost({
         userId: user.id,
         postId: post.id,
         content: "content",
@@ -488,7 +349,7 @@ describe("PostService", () => {
       });
 
       // When: 글 업데이트 시도
-      const result = postService.updatePost({
+      const result = postMutationService.updatePost({
         userId: user.id,
         postId: post.id,
         content: "content",
@@ -508,7 +369,7 @@ describe("PostService", () => {
       const post = await saveTestPost(postRepository, user.id);
 
       // When: liked를 true로 지정하여 요청
-      await postService.setPostLike(user.id, post.id, true);
+      await postMutationService.setPostLike(user.id, post.id, true);
 
       // Then: post like entity 생성, likeCount 업데이트
       const like = await postLikeRepository.findOneBy({
@@ -536,7 +397,7 @@ describe("PostService", () => {
       );
 
       // When: liked를 true로 지정하여 요청
-      await postService.setPostLike(user.id, post.id, true);
+      await postMutationService.setPostLike(user.id, post.id, true);
 
       // Then: post like entity 유지, likeCount 유지
       const like = await postLikeRepository.findOne({
@@ -561,7 +422,7 @@ describe("PostService", () => {
       );
 
       // When: liked를 false로 지정하여 요청
-      await postService.setPostLike(user.id, post.id, false);
+      await postMutationService.setPostLike(user.id, post.id, false);
 
       // Then: post like entity 삭제, likeCount 업데이트
       const like = await postLikeRepository.findOne({
@@ -578,7 +439,7 @@ describe("PostService", () => {
       const post = await saveTestPost(postRepository, user.id);
 
       // When: liked를 false로 지정하여 요청
-      await postService.setPostLike(user.id, post.id, false);
+      await postMutationService.setPostLike(user.id, post.id, false);
 
       // Then: post like entity 없음 유지, likeCount 유지
       const like = await postLikeRepository.findOne({
@@ -594,7 +455,7 @@ describe("PostService", () => {
       const user = await saveTestUser(userRepository);
 
       // When: liked 요청 시도
-      const result = postService.setPostLike(user.id, 0, true);
+      const result = postMutationService.setPostLike(user.id, 0, true);
 
       // Then: throws handled error
       await expect(result).rejects.toThrow(PostNotFoundError);
@@ -612,7 +473,7 @@ describe("PostService", () => {
       fakeS3.put(attachment.s3Key);
 
       // When: 글 삭제 시도
-      await postService.deletePost(user.id, post.id);
+      await postMutationService.deletePost(user.id, post.id);
 
       // Then: 글 및 attachments 삭제됨
       expect(await postRepository.findOneBy({ id: post.id })).toBeNull();
@@ -627,7 +488,7 @@ describe("PostService", () => {
       const user = await saveTestUser(userRepository);
 
       // When: 글 삭제 시도
-      const result = postService.deletePost(user.id, 0);
+      const result = postMutationService.deletePost(user.id, 0);
 
       // Then: throws handled error
       await expect(result).rejects.toThrow(PostNotFoundError);
@@ -640,7 +501,7 @@ describe("PostService", () => {
       const post = await saveTestPost(postRepository, otherUser.id);
 
       // When: 글 삭제 시도
-      const result = postService.deletePost(user.id, post.id);
+      const result = postMutationService.deletePost(user.id, post.id);
 
       // Then: throws handled error
       await expect(result).rejects.toThrow(NotPostAuthorError);
