+++ apps/backend/src/post/post-mutation.service.ts
@@ -8,7 +8,6 @@ import { S3StorageService } from "src/s3-storage/s3-storage.service";
 import {
   AttachmentAlreadyAssociatedError,
   AttachmentNotFoundError,
-  InvalidCursorError,
   NotPostAuthorError,
   PostNotFoundError,
 } from "./post.errors";
@@ -21,21 +20,6 @@ export type CreatePostOptions = {
   attachmentIds: number[];
 };
 
-export type ListPostsOrderBy = "createdAt";
-
-export type ListPostsOptions = {
-  cursor?: string;
-  limit: number;
-  category?: PostCategory;
-  orderBy: ListPostsOrderBy;
-  orderDirection: "asc" | "desc";
-};
-
-export type ListPostsResult = {
-  posts: Post[];
-  nextCursor?: string;
-};
-
 export type UpdatePostOptions = {
   userId: number;
   postId: number;
@@ -44,45 +28,11 @@ export type UpdatePostOptions = {
   attachmentIds: number[];
 };
 
-type CursorPayload = {
-  category: PostCategory | null;
-  orderBy: ListPostsOrderBy;
-  orderDirection: "asc" | "desc";
-  cursorValue: string;
-  cursorId: number;
-};
-
-function encodeCursor(payload: CursorPayload): string {
-  return Buffer.from(JSON.stringify(payload)).toString("base64url");
-}
-
-function decodeCursor(cursor: string): CursorPayload {
-  return JSON.parse(Buffer.from(cursor, "base64url").toString("utf8"));
-}
-
-// 2026-03-25 code smell....
-const orderByMap: Record<
-  ListPostsOrderBy,
-  {
-    path: string;
-    column: string;
-    cast: string;
-    getValue: (post: Post) => string;
-  }
-> = {
-  createdAt: {
-    path: "post.createdAt",
-    column: 'extract(epoch FROM "post"."created_at")::int',
-    cast: "extract(epoch FROM :cursorValue)::int",
-    getValue: (post) => post.createdAt.toISOString(),
-  },
-};
-
 /**
- * PostController와 1:1로 대응되는, post 관련 orchestration logic을 담당한다.
+ * controller에서 사용하기 위한 글의 mutation을 담당한다.
  */
 @Injectable()
-export class PostService {
+export class PostMutationService {
   constructor(
     @InjectRepository(Post)
     private readonly postRepository: Repository<Post>,
@@ -118,86 +68,6 @@ export class PostService {
       }),
     );
 
-    return post;
-  }
-
-  /*
-   * pagination option을 적용하여 글 목록을 조회한다.
-   * @returns 글 목록, attachment URL, cursor
-   */
-  async listPosts({
-    limit,
-    orderBy,
-    orderDirection,
-    category,
-    cursor,
-  }: ListPostsOptions): Promise<ListPostsResult> {
-    const decodedCursor = cursor && decodeCursor(cursor);
-
-    if (
-      decodedCursor &&
-      (category != decodedCursor.category || // != for null-undefined checks
-        orderBy !== decodedCursor.orderBy ||
-        orderDirection !== decodedCursor.orderDirection)
-    ) {
-      throw new InvalidCursorError();
-    }
-
-    const sqlDirection = orderDirection === "asc" ? "ASC" : "DESC";
-    const cursorOp = orderDirection === "asc" ? ">" : "<";
-    const orderEntry = orderByMap[orderBy];
-
-    const qb = this.postRepository.createQueryBuilder("post");
-
-    if (category) {
-      qb.where("post.category = :category", { category });
-    }
-
-    if (decodedCursor) {
-      qb.andWhere(
-        [
-          `(${orderEntry.column}, "post"."id")`,
-          cursorOp,
-          `(${orderEntry.cast}, :cursorId::int)`,
-        ].join(""),
-        {
-          cursorValue: decodedCursor.cursorValue,
-          cursorId: decodedCursor.cursorId,
-        },
-      );
-    }
-
-    const posts = await qb
-      .orderBy(orderEntry.path, sqlDirection)
-      .addOrderBy("post.id", sqlDirection)
-      .limit(limit)
-      .getMany();
-
-    const lastPost = posts.at(-1)!;
-    return {
-      posts,
-      nextCursor:
-        posts.length === limit
-          ? encodeCursor({
-              orderBy,
-              orderDirection,
-              category: category ?? null,
-              cursorId: lastPost.id,
-              cursorValue: orderEntry.getValue(lastPost),
-            })
-          : undefined,
-    };
-  }
-
-  /**
-   * postId에 대응하는 글 하나를 조회한다.
-   * @returns 해당 글, attachment URL
-   */
-  async getPost(postId: number): Promise<Post> {
-    const post = await this.postRepository.findOneBy({ id: postId });
-    if (!post) {
-      throw new PostNotFoundError();
-    }
     return post;
   }
 
@@ -251,7 +121,7 @@ export class PostService {
     });
 
     await this.s3StorageService.deleteMany(
-      post.attachments.map((a) => a.s3Key),
+      deletingAttachments.map((a) => a.s3Key),
     );
   }
 
@@ -305,7 +175,7 @@ export class PostService {
   }
 
   /**
-   * userId에 대응하는 사용자가 글의 작성자인 경우, 글을 삭제한다.2
+   * userId에 대응하는 사용자가 글의 작성자인 경우, 글을 삭제한다.
    */
   async deletePost(userId: number, postId: number): Promise<void> {
     const post = await this.postRepository.findOne({
